+++ Importia.SDK/Tests/a3innuva.TAA.Migration.SDK.Implementations.Tests/Validations/FixedAssetValidationsTests.cs
@@ -247,8 +247,8 @@ namespace a3innuva.TAA.Migration.SDK.Implementations.Tests
             public bool RetirementIsExempt { get; set; }
             public DateTime AssetStartDate { get; set; }
             public decimal? DepreciationResidualValue { get; set; }
-            public int DepreciationYears { get; set; }
-            public int DepreciationFiscalYears { get; set; }
+            public decimal DepreciationYears { get; set; }
+            public decimal DepreciationFiscalYears { get; set; }
             public IEnumerable<IFixedAssetDepreciationQuota> DepreciationQuotas { get; set; }
 
             public string Identity() => $"{Id}";
